refactor(directory): update directory
