+++ package.json
@@ -35,18 +35,19 @@
     "clean": "rimraf ./lib",
     "cli": "node ./lib/cli.js",
     "commitmsg": "commitlint -e $GIT_PARAMS",
+    "cruiser:image": "depcruise  --exclude \"^node_modules\" --output-type dot src/server | dot -T png > docs/dependencygraph.png",
     "develop": "yarn run build --watch",
     "develop:server": "chmod +x ./lib/cli.js && ./lib/cli.js example -Do --port 9000",
     "format": "prettier --config .prettierrc --write src/*.{ts,tsx}",
     "lint": "tslint -c tslint.json -p tsconfig.json",
     "lint:fix": "yarn run lint --fix",
+    "prettier": "prettier --config .prettierrc --write src/*.{ts,tsx}",
     "start": "run-p *:watch",
     "test": "yarn run test:cruiser && yarn run test:jest",
     "test:ci": "yarn run test && codecov",
     "test:cruiser": "dependency-cruiser --validate .dependency-cruiser.json src",
     "test:jest": "jest -c ./jest.config.json",
     "test:watch": "yarn run test:jest --watch",
-    "prettier": "prettier --config .prettierrc --write src/*.{ts,tsx}",
     "version_up:major": "changelog -M && git add CHANGELOG.md && git commit -m \"chore(change-log): updated CHANGELOG.md\" && npm version major",
     "version_up:minor": "changelog -m && git add CHANGELOG.md && git commit -m \"chore(change-log): updated CHANGELOG.md\" && npm version minor",
     "version_up:patch": "changelog -p && git add CHANGELOG.md && git commit -m \"chore(change-log): updated CHANGELOG.md\" && npm version patch"
@@ -89,7 +90,7 @@
     "@types/read-pkg-up": "^3.0.1",
     "@types/update-notifier": "^2.5.0",
     "codecov": "^3.1.0",
-    "dependency-cruiser": "^4.5.0",
+    "dependency-cruiser": "^4.8.0",
     "generate-changelog": "^1.7.1",
     "husky": "^1.1.3",
     "jest": "^23.5.0",
