+++ src/generator.ts
@@ -1,8 +1,8 @@
 import { Options } from "@rocu/cli";
 import { RenderedStaticPage, Source } from "@rocu/page";
-import { notifyLog } from "../logger";
-import { render } from "../renderer";
-import { getData } from "../structure";
+import { notifyLog } from "./logger";
+import { render } from "./renderer";
+import { getData } from "./structure/getPage";
 
 export const generateStatic = async (source: Source, opts: Options): Promise<RenderedStaticPage[]> => {
   return render(source, opts);
