+++ src/index.ts
@@ -1,6 +1,3 @@
-export * from "./transformer";
 export { generateStaticPage } from "./generator";
 export { notifyLog } from "./logger";
-export { render } from "./renderer";
 export { server } from "./server";
-export { getData } from "./structure";
