--- src/structure/index.ts
@@ -1,1 +1,0 @@
-export { getData } from "./getPage";
