--- src/transformer/index.tsx
@@ -1,1 +1,0 @@
-export { transformRawStringToHtml } from "./converter";
